[refactor/#63] activity 페이지 컴포넌트 구조 개선
